+++ backend/src/main/java/it/unipi/dii/aide/lsmd/readrumble/book/LightBookDTO.java
@@ -1,15 +1,15 @@
 package it.unipi.dii.aide.lsmd.readrumble.book;
 
 public class LightBookDTO {
-    private int id;
+    private long id;
     private String title;
 
-    public LightBookDTO(int id, String title) {
+    public LightBookDTO(long id, String title) {
         this.id = id;
         this.title = title;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
